pull out calculate_adjacency_matrix method from Residence, move to utils
