+++ solvation_analysis/residence.py
@@ -25,6 +25,7 @@ from statsmodels.tsa.stattools import acovf
 from scipy.optimize import curve_fit
 
 from solvation_analysis._column_names import *
+from solvation_analysis._utils import calculate_adjacency_dataframe
 
 
 class Residence:
@@ -130,7 +131,7 @@ class Residence:
         frame_solute_index = pd.MultiIndex.from_tuples(unique_indices, names=partial_index.names)
         auto_covariance_dict = {}
         for res_name, res_solvation_data in self.solvation_data.groupby([SOLVENT]):
-            adjacency_mini = Residence.calculate_adjacency_dataframe(res_solvation_data)
+            adjacency_mini = calculate_adjacency_dataframe(res_solvation_data)
             adjacency_df = adjacency_mini.reindex(frame_solute_index, fill_value=0)
             auto_covariance = Residence._calculate_auto_covariance(adjacency_df)
             # normalize
@@ -252,28 +253,4 @@ class Residence:
             auto_covariances.append(auto_covariance_df.values)
         auto_covariance = np.mean(np.concatenate(auto_covariances, axis=1), axis=1)
         return auto_covariance
-
-    @staticmethod
-    def calculate_adjacency_dataframe(solvation_data):
-        """
-        Calculate a frame-by-frame adjacency matrix from the solvation data.
-
-        This will calculate the adjacency matrix of the solute and all possible
-        solvents. It will maintain an index of ["frame", "solute_atom", "solvent"]
-        where each "frame" is a sparse adjacency matrix between solvated atom ix
-        and residue ix.
-
-        Parameters
-        ----------
-        solvation_data : pd.DataFrame
-            the solvation_data from a Solute.
-
-        Returns
-        -------
-        adjacency_df : pandas.DataFrame
-        """
-        # generate an adjacency matrix from the solvation data
-        adjacency_group = solvation_data.groupby([FRAME, SOLUTE_IX, SOLVENT_IX])
-        adjacency_df = adjacency_group[DISTANCE].count().unstack(fill_value=0)
-        return adjacency_df
